Extract mutations_recording
